# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "UniConnect API",
    "version": "1.0.0",
    "description": "API REST de la red social universitaria UniConnect"
  },
  "servers": [
    {
      "url": "https://uniconnect-backend.fly.dev",
      "description": "Producción"
    },
    {
      "url": "http://localhost:3000",
      "description": "Desarrollo local"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Usuario": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "nombre": {
            "type": "string"
          },
          "apellido": {
            "type": "string"
          },
          "correo": {
            "type": "string",
            "format": "email"
          },
          "carrera": {
            "type": "string"
          },
          "semestre": {
            "type": "integer"
          }
        }
      },
...
      "Grupo": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "nombre": {
            "type": "string"
          },
          "materiaId": {
            "type": "string"
          },
          "materia": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string"
              },
              "nombre": {
                "type": "string"
              }
            }
          },
          "creadorId": {
            "type": "string"
          },
          "administradorId": {
            "type": "string"
          },
          "cantidadMiembros": {
            "type": "integer"
          },
          "miembros": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Usuario"
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "Mensaje": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "contenido": {
            "type": "string"
          },
          "remitenteId": {
            "type": "string"
          },
          "remitenteNombre": {
            "type": "string"
          },
          "destinatarioId": {
            "type": "string"
          },
          "grupoId": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/grupos": {
      "post": {
        "summary": "Crear un nuevo grupo de estudio",
        "tags": [
          "Grupos"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "nombre",
                  "materiaId"
                ],
                "properties": {
                  "nombre": {
                    "type": "string"
                  },
                  "materiaId": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Grupo creado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Grupo"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Listar grupos del usuario autenticado",
        "tags": [
          "Grupos"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de grupos del usuario",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Grupo"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/grupos/buscar": {
      "get": {
        "summary": "Buscar grupos por nombre o materia",
        "tags": [
          "Grupos"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Grupos encontrados"
          }
        }
      }
    },
    "/api/grupos/disponibles": {
      "get": {
        "summary": "Listar grupos disponibles para unirse (mismas materias del usuario)",
        "tags": [
          "Grupos"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Grupos disponibles"
          }
        }
      }
    },
    "/api/grupos/mis-solicitudes": {
      "get": {
        "summary": "Listar solicitudes de ingreso enviadas por el usuario",
        "tags": [
          "Grupos"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Solicitudes enviadas"
          }
        }
      }
    },
    "/api/grupos/{id}": {
      "get": {
        "summary": "Obtener detalle de un grupo",
        "tags": [
          "Grupos"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Detalle del grupo",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Grupo"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/grupos/{id}/miembros": {
      "get": {
        "summary": "Listar miembros de un grupo",
        "tags": [
          "Grupos"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de miembros"
          }
        }
      },
      "post": {
        "summary": "Agregar un miembro al grupo (solo administrador)",
        "tags": [
          "Grupos"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "usuarioId"
                ],
                "properties": {
                  "usuarioId": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Miembro agregado"
          }
        }
      }
    },
    "/api/grupos/{id}/solicitar-ingreso": {
      "post": {
        "summary": "Enviar solicitud de ingreso a un grupo",
        "tags": [
          "Grupos"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Solicitud enviada"
          }
        }
      }
    },
    "/api/grupos/{id}/solicitudes": {
      "get": {
        "summary": "Listar solicitudes de ingreso al grupo (solo administrador)",
        "tags": [
          "Grupos"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de solicitudes pendientes"
          }
        }
      }
    },
    "/api/grupos/{id}/solicitudes/{solicitudId}/aprobar": {
      "patch": {
        "summary": "Aprobar solicitud de ingreso",
        "tags": [
          "Grupos"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "solicitudId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Solicitud aprobada"
          }
        }
      }
    },
    "/api/grupos/{id}/solicitudes/{solicitudId}/rechazar": {
      "patch": {
        "summary": "Rechazar solicitud de ingreso",
        "tags": [
          "Grupos"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "solicitudId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Solicitud rechazada"
          }
        }
      }
    },
    "/api/grupos/{id}/abandonar": {
      "delete": {
        "summary": "Abandonar un grupo",
        "tags": [
          "Grupos"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Grupo abandonado exitosamente"
          },
          "409": {
            "description": "El administrador debe transferir el rol antes de abandonar"
          }
        }
      }
    },
    "/api/grupos/{id}/archivos": {
      "post": {
        "summary": "Subir archivo PDF al grupo",
        "tags": [
          "Grupos"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "archivo": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Archivo subido a Cloudinary"
          }
        }
      },
      "get": {
        "summary": "Listar archivos del grupo",
        "tags": [
          "Grupos"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de archivos"
          }
        }
      }
    },
    "/api/grupos/{id}/archivos/{archivoId}/descargar": {
      "get": {
        "summary": "Descargar un archivo del grupo",
        "tags": [
          "Grupos"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "archivoId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "URL de descarga del archivo"
          }
        }
      }
    },
    "/api/grupos/{id}/administrador/iniciar": {
      "post": {
        "summary": "El admin nomina un candidato para la transferencia de administración",
        "tags": [
          "Grupos"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "candidatoId"
                ],
                "properties": {
                  "candidatoId": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Transferencia iniciada, candidato notificado"
          }
        }
      }
    },
    "/api/grupos/{id}/administrador/aceptar": {
      "post": {
        "summary": "El candidato acepta la administración del grupo",
        "tags": [
          "Grupos"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Transferencia aceptada, admin actualizado"
          }
        }
      }
    },
    "/api/grupos/{id}/administrador/rechazar": {
      "post": {
        "summary": "El candidato rechaza la administración del grupo",
        "tags": [
          "Grupos"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Transferencia rechazada, grupo vuelve a ACTIVO"
          }
        }
      }
    },
    "/api/grupos/{id}/administrador/cancelar": {
      "delete": {
        "summary": "El admin cancela la transferencia en curso",
        "tags": [
          "Grupos"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Transferencia cancelada, grupo vuelve a ACTIVO"
          }
        }
      }
    },
    "/api/grupos/{id}/mensajes": {
      "post": {
        "summary": "Enviar mensaje al chat del grupo",
        "tags": [
          "Grupos"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "contenido"
                ],
                "properties": {
                  "contenido": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Mensaje enviado"
          }
        }
      },
      "get": {
        "summary": "Obtener historial de mensajes del grupo",
        "tags": [
          "Grupos"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 50
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Historial de mensajes"
          }
        }
      }
    },
...
    "/api/mensajes/grupos": {
      "post": {
        "summary": "Enviar mensaje a un grupo",
        "tags": [
          "Mensajes"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "grupoId",
                  "contenido"
                ],
                "properties": {
                  "grupoId": {
                    "type": "string"
                  },
                  "contenido": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Mensaje enviado al grupo",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Mensaje"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/mensajes/grupos/{grupoId}": {
      "get": {
        "summary": "Obtener historial de mensajes de un grupo",
        "tags": [
          "Mensajes"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "grupoId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 50
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Historial de mensajes del grupo",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Mensaje"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}